+++ src/main/java/webserver/controller/Controller.java
@@ -1,0 +1,8 @@
+package webserver.controller;
+
+import webserver.request.HttpRequest;
+import webserver.response.HttpResponse;
+
+public interface Controller {
+    HttpResponse service(HttpRequest request);
+}
